requirements.txt
absl-py==0.8.1
audioread==2.1.8
cffi==1.12.3
decorator==4.4.0
future==0.18.1
grpcio==1.24.1
joblib==0.14.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
numba==0.45.1
numpy==1.17.2
pandas==0.25.1
Pillow==6.2.0
protobuf==3.10.0
pycparser==2.19
python-dateutil==2.8.0
pytorch-land==0.1.6
pytz==2019.3
resampy==0.2.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
SoundFile==0.10.2
tensorboard==2.0.0
torch==1.2.0
torchvision==0.4.0
Werkzeug==0.16.0
